# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.AnalyticsOverview:
    properties:
      this_month:
        type: integer
      this_week:
        type: integer
      today:
        type: integer
      total_clicks:
        type: integer
    type: object
...
  models.BrowserAnalytics:
    properties:
      browser:
        type: string
      count:
        type: integer
    type: object
...
  models.DailyClick:
    properties:
      clicks:
        type: integer
      date:
        type: string
    type: object
  models.DeviceAnalytics:
    properties:
      count:
        type: integer
      device:
        type: string
    type: object
...
  models.PaginatedURLResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.URL'
        type: array
      meta:
        $ref: '#/definitions/models.PaginationMeta'
    type: object
  models.PaginatedUserResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.User'
        type: array
      meta:
        $ref: '#/definitions/models.PaginationMeta'
    type: object
  models.PaginationMeta:
    properties:
      limit:
        example: 50
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 150
        type: integer
      total_pages:
        example: 3
        type: integer
    type: object
  models.RecentVisit:
    properties:
      browser:
        type: string
      created_at:
        type: string
      device:
        type: string
      ip_address:
        type: string
      referer:
        type: string
    type: object
...
  models.URL:
    properties:
      clicks:
        type: integer
      created_at:
        type: string
      expiry:
        type: string
      id:
        description: go for uuid later
        type: integer
      long_url:
        type: string
      short_url:
        type: string
      user_id:
        type: integer
    required:
    - long_url
    - short_url
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: URL shortener with Redis caching, sliding-window limits, buffered analytics
    and Prometheus metrics.
  title: URL Shortener
  version: "1.0"
paths:
...
  /api/v1/admin/users:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedUserResponse'
      security:
      - Bearer: []
      summary: List users as an administrator
      tags:
      - admin
  /api/v1/admin/users/{userid}/urls:
    get:
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedURLResponse'
      security:
      - Bearer: []
      summary: List a user's URLs as an administrator
      tags:
      - admin
  /api/v1/analytics/{id}/browser:
    get:
      parameters:
      - description: Owned URL ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BrowserAnalytics'
            type: array
      security:
      - Bearer: []
      summary: Get sampled visit counts by browser
      tags:
      - legacy analytics
  /api/v1/analytics/{id}/daily:
    get:
      parameters:
      - description: Owned URL ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DailyClick'
            type: array
      security:
      - Bearer: []
      summary: Get daily sampled visit counts in UTC
      tags:
      - legacy analytics
  /api/v1/analytics/{id}/device:
    get:
      parameters:
      - description: Owned URL ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DeviceAnalytics'
            type: array
      security:
      - Bearer: []
      summary: Get sampled visit counts by device
      tags:
      - legacy analytics
  /api/v1/analytics/{id}/overview:
    get:
      parameters:
      - description: URL ID (base, premium or admin role required)
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AnalyticsOverview'
      security:
      - Bearer: []
      summary: Get an owned URL's sampled visit overview
      tags:
      - legacy analytics
  /api/v1/analytics/{id}/recent:
    get:
      parameters:
      - description: Owned URL ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RecentVisit'
            type: array
      security:
      - Bearer: []
      summary: Get recent sampled visits
      tags:
      - legacy analytics
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"