# backend/internal/handler/analytics_controller.go
package controllers

import (
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"
	"github.com/gottatouchsomegrass/url/internal/model"
	"github.com/gottatouchsomegrass/url/internal/service"
)

type AnalyticsController struct {
	Service *services.AnalyticsService
}

func NewAnalyticsController(s *services.AnalyticsService) *AnalyticsController {
	return &AnalyticsController{
		Service: s,
	}
}

func parseURLID(c *gin.Context) (int64, error) {
	urlID, err := strconv.ParseInt(c.Param("id"), 10, 64)
	if err != nil {
		return 0, err
	}
	return urlID, nil
}

func handleServiceError(c *gin.Context, err error) {
	switch err.Error() {
	case "url not found":
		c.JSON(http.StatusNotFound, models.HTTPError{Error: err.Error()})
	case "forbidden":
		c.JSON(http.StatusForbidden, models.HTTPError{Error: err.Error()})
	default:
		c.JSON(http.StatusInternalServerError, models.HTTPDetailsError{
			Error:   "internal server error",
			Details: err.Error(),
		})
	}
}

// GetAnalyticsOverview godoc
// @Summary Get an owned URL's sampled visit overview
// @Tags legacy analytics
// @Produce json
// @Security Bearer
// @Param id path int true "URL ID (base, premium or admin role required)"
// @Success 200 {object} models.AnalyticsOverview
// @Router /api/v1/analytics/{id}/overview [get]
func (ac *AnalyticsController) GetAnalyticsOverview(c *gin.Context) {
	ctx := c.Request.Context()
	userID := c.MustGet("userID").(int64)

	urlID, err := parseURLID(c)
	if err != nil {
		c.JSON(http.StatusBadRequest, models.HTTPError{Error: "invalid url id"})
		return
	}

	overview, err := ac.Service.GetAnalyticsOverview(ctx, urlID, userID)
	if err != nil {
		handleServiceError(c, err)
		return
	}

	c.JSON(http.StatusOK, overview)
}

// GetDailyClicks godoc
// @Summary Get daily sampled visit counts in UTC
// @Tags legacy analytics
// @Produce json
// @Security Bearer
// @Param id path int true "Owned URL ID"
// @Success 200 {array} models.DailyClick
// @Router /api/v1/analytics/{id}/daily [get]
func (ac *AnalyticsController) GetDailyClicks(c *gin.Context) {
	ctx := c.Request.Context()
	userID := c.MustGet("userID").(int64)

	urlID, err := parseURLID(c)
	if err != nil {
		c.JSON(http.StatusBadRequest, models.HTTPError{Error: "invalid url id"})
		return
	}

	clicks, err := ac.Service.GetDailyClicks(ctx, urlID, userID)
	if err != nil {
		handleServiceError(c, err)
		return
	}

	c.JSON(http.StatusOK, clicks)
}

// GetRecentVisits godoc
// @Summary Get recent sampled visits
// @Tags legacy analytics
// @Produce json
// @Security Bearer
// @Param id path int true "Owned URL ID"
// @Success 200 {array} models.RecentVisit
// @Router /api/v1/analytics/{id}/recent [get]
func (ac *AnalyticsController) GetRecentVisits(c *gin.Context) {
	ctx := c.Request.Context()
	userID := c.MustGet("userID").(int64)

	urlID, err := parseURLID(c)
	if err != nil {
		c.JSON(http.StatusBadRequest, models.HTTPError{Error: "invalid url id"})
		return
	}

	visits, err := ac.Service.GetRecentVisits(ctx, urlID, userID)
	if err != nil {
		handleServiceError(c, err)
		return
	}

	c.JSON(http.StatusOK, visits)
}

// GetBrowserAnalytics godoc
// @Summary Get sampled visit counts by browser
// @Tags legacy analytics
// @Produce json
// @Security Bearer
// @Param id path int true "Owned URL ID"
// @Success 200 {array} models.BrowserAnalytics
// @Router /api/v1/analytics/{id}/browser [get]
func (ac *AnalyticsController) GetBrowserAnalytics(c *gin.Context) {
	ctx := c.Request.Context()
	userID := c.MustGet("userID").(int64)

	urlID, err := parseURLID(c)
	if err != nil {
		c.JSON(http.StatusBadRequest, models.HTTPError{Error: "invalid url id"})
		return
	}

	browserStats, err := ac.Service.GetBrowserAnalytics(ctx, urlID, userID)
	if err != nil {
		handleServiceError(c, err)
		return
	}

	c.JSON(http.StatusOK, browserStats)
}

// GetDeviceAnalytics godoc
// @Summary Get sampled visit counts by device
// @Tags legacy analytics
// @Produce json
// @Security Bearer
// @Param id path int true "Owned URL ID"
// @Success 200 {array} models.DeviceAnalytics
// @Router /api/v1/analytics/{id}/device [get]
func (ac *AnalyticsController) GetDeviceAnalytics(c *gin.Context) {
	ctx := c.Request.Context()
	userID := c.MustGet("userID").(int64)

	urlID, err := parseURLID(c)
	if err != nil {
		c.JSON(http.StatusBadRequest, models.HTTPError{Error: "invalid url id"})
		return
	}

	deviceStats, err := ac.Service.GetDeviceAnalytics(ctx, urlID, userID)
	if err != nil {
		handleServiceError(c, err)
		return
	}

	c.JSON(http.StatusOK, deviceStats)
}